+++ api-testcases/test-errhandler.c
@@ -34,7 +34,7 @@ int test_1(duk_context *ctx) {
 	duk_push_error_object(ctx, DUK_ERR_RANGE_ERROR, "range error: %d", 123);
 	duk_throw(ctx);
 
-	printf("final top: %d\n", duk_get_top(ctx));
+	printf("final top: %ld\n", (long) duk_get_top(ctx));
 	return 0;
 }
 
@@ -49,7 +49,7 @@ int test_2(duk_context *ctx) {
 
 	duk_error(ctx, 1234567, "arbitrary error code");
 
-	printf("final top: %d\n", duk_get_top(ctx));
+	printf("final top: %ld\n", (long) duk_get_top(ctx));
 	return 0;
 }
 
@@ -66,7 +66,7 @@ int test_3(duk_context *ctx) {
 
 	duk_error(ctx, DUK_ERR_TYPE_ERROR, NULL);
 
-	printf("final top: %d\n", duk_get_top(ctx));
+	printf("final top: %ld\n", (long) duk_get_top(ctx));
 	return 0;
 }
 
@@ -83,7 +83,7 @@ int test_4(duk_context *ctx) {
 	printf("string coerced: %s\n", duk_to_string(ctx, -1));
 	duk_pop(ctx);
 
-	printf("final top: %d\n", duk_get_top(ctx));
+	printf("final top: %ld\n", (long) duk_get_top(ctx));
 	return 0;
 }
 
@@ -97,7 +97,7 @@ int test_5(duk_context *ctx) {
 
 	duk_error(ctx, 1234567, "arbitrary error code");
 
-	printf("final top: %d\n", duk_get_top(ctx));
+	printf("final top: %ld\n", (long) duk_get_top(ctx));
 	return 0;
 }
